backend/apps/users/serializers/auth_serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User, Group
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer


class RegisterSerializer(serializers.ModelSerializer):
    role = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('username', 'email', 'password', 'role')
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        role = validated_data.pop('role', None)
        user = User.objects.create_user(**validated_data)
        if role:
            group, _ = Group.objects.get_or_create(name=role)
            user.groups.add(group)
        return user

class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
    def validate(self, attrs):
        data = super().validate(attrs)

        # Add the user's role (group name) to the response
        user = self.user
        roles = user.groups.values_list('name', flat=True)
        data['role'] = roles[0] if roles else 'NoRole'

        return data