backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==42.0.8
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
